// src/modules/food/interfaces/NutrientProfile.ts
export interface NutrientProfile {
  // Macronutrients
  protein: number;
  fat: number;
  carbs: number;
  fiber: number;
  sugar: number;
  sodium: number;
  
  // Vitamins
  vitaminA: number;
  vitaminB1: number;
  vitaminB2: number;
  vitaminB3: number;
  vitaminB5: number;
  vitaminB6: number;
  vitaminB7: number;
  vitaminB9: number;
  vitaminB12: number;
  vitaminC: number;
  vitaminD: number;
  vitaminE: number;
  vitaminK: number;
  
  // Minerals
  calcium: number;
  iron: number;
  magnesium: number;
  potassium: number;
  zinc: number;
}